Enforce callbacks for ext_* functions instead of linked implementations.
This is needed to show a single API to user applications. Before this
patch, one could use the USE_STRUCT_CALLBACKS macro to enable/disable
callback support. This meant that distributions will have to choose
one of two and applications that use the other will break. This patch
breaks compatibility with applications that link to a copy of
libyahoo2 that has not been compiled with USE_STRUCT_CALLBACKS
defined. Applications that already define this should work OK.

Another side effect of this fix is that it makes autoconf checks for
libyahoo2 easier.